# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Motolegacy",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://KidziBack.squaretech.tech/motolegacy/api"
        },
        {
            "url": "http://127.0.0.1:8000/motolegacy/api"
        }
    ],
    "paths": {
...
        "/admin/categories": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "List all categories",
                "operationId": "0cf6e89e69cabba5e72db6323e787007",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Create a new category (Admin only)",
                "operationId": "a533ff285fab1552a0e830b3d4e76d82",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/categories/{id}": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Get category details",
                "operationId": "4f4954eaefabb4b722c23de0bb0c9c4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Update a category (Admin only)",
                "operationId": "9e23a7f4760a9b72457b181d53a07ec5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PATCH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Delete a category (Admin only)",
                "operationId": "0ee11e2aa6e87c46d4a4a8c156605743",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/admin/roles": {
            "get": {
                "tags": [
                    "Admin Roles"
                ],
                "description": "get all roles",
                "operationId": "a7d4464f8781e2ae359e12db6d703250",
                "parameters": [
                    {
                        "name": "with_paginate",
                        "in": "query",
                        "description": "Enable pagination (0 = false, 1 = true)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}